Fix for two things:
(1) General: Using --enable-srvloc results in compile-time errors when
libslp.a is not installed in a standard location searched by the linker.
This is due to the missing third argument in AC_CHECK_LIB(slp,
SLPOpen, , AC_MSG_ERROR([SLP installation not found])) on line  32 in
macros/srvloc.m4: This sets LIBS="-lslp $LIBS" globally in configure,
instead of just setting SLP_LIBS and SLP_CFLAGS.

(2) Tru64-specific: In libatalk/utils/logger.c, the function vsnprintf()
is used. Up to and including version 4.0F, Tru64 doesn't have this
function in its standard C library, although the DEC C compiler has a
prototype for it. Instead, vsnprintf() is contained in libdb.a (not to
be confused with libdb3.a), which must be given as a separate flag to
the linker.

Submitted by Burkhard Schmidt <bs@cpfs.mpg.de>.