Introduce Conncurrent Data Store (CDB) support to CNID.  This is phase 1. In this phase, CDB is a configure option, and is either enabled or disabled for all volumes at compile time.  By default, CDB is disabled in favor of transactional data store.
CDB has some advantages.  It doesn't use transactions or logging, thus it's
less likely to be corrupted if the server dies prematurely.  It also
doesn't eat up disk space with transaction logs.  The disadvantage is that
it's designed for environments with many readers and one writer.  This means
that many afpd clients could block while waiting for a write lock on
the database.  That's why testing is in order.

In phase 2, CDB is be configurable on a per-volume basis.

Phase 3 (if testing of phase 1 and 2 is good) will be to enable CDB as the
default CNID data store scheme.

To enable CDB, simply specify the configure argument --with-cdb when
configuring with CNID DID support.

Submitted by:	Dan Wilga <dwilga@mtholyoke.edu>