Change status_directorynames to use keytab when constructing names
v3: Forgot to pass the address of 'data' to append_directoryname (since it adjusts it by the length of the principal)
v2: Add append_directory name
v2: Construct a directory name even if Kerberos support is disabled

When constructing the DirectoryNames portion of the response, we check if uams_gss is loaded. If it isn't, we don't bother with constructing a name. If uams_gss *is* loaded, we need to check the options k5service, fqdn and k5realm. If they are set, we construct a service principal name using them, and ensure that principal is present in afpd's keytab. Otherwise, we use the principal from the first entry in the keytab. After we have a valid principal, we convert it to a directory name, and use it for the response.

AFPSRVRINFO_SRVRDIR implies that we can authenticate via Kerberos, so only enable that if uams_gss is loaded.

Signed-off-by: Jamie Gilbertson <jamie.gilbertson@luniv.ca>