Fix dsi->in_write flag/counter
The new/tuned sendfile using code resulted in in_write being
decremented to -1 because dsi_readdone was called (which
decremnts it by 1) althoug it wasn't incremented before. The fix
is to not call dsi_readdone() at all in the sendfile case as
we didn't call the counterpart dsi_readinit() beforehand.