Ignore db stamp when reading CNID from adouble file
The db stamp (which is the ctime of the cnid2.db file stored in  the
db at creation time) may change when running `dbd -f` or when deleting
the CNID db manually. But we don't want to reassing CNIDs in this case
(changed stamp) we really want to keep it.