Increase IO size when sendfile() is not used
If sendfile() support is disabled, the current code results in an
IO size of 8k for transferring requested data to the client. Eg:
- client request to read DSI quantum size bytes (~250k) from a file
- afpd process loops read(8k, file) -> send(8k, client) until all
  250k have been transferred
This should be modified to read the whole 250k in one swoop from the
file to a buffer and then send the whole buffer in one swoop to the
client.

From FR #76.