Remove size check that required a costly fstat()
In case the client request count plus offset is large then the file,
we're supposed to send the available data and set AFP error to
AFPERR_EOF.
No need to use an extra fstat() for that, we can also simply check
the whether the amount of data read from the file matches the
requested amount.

From FR #78.