New boolean volume option "delete veto files"
If this option is set to yes, then Netatalk will attempt to
recursively delete any vetoed files and directories.

Remove obselete code that tries to delete dangling symlinks, this
is not necessary anymore since Netatalk support UNIX symlinks.

Modify behaviour of the VFS function deletecurdir of the "ea" VFS
module so it deletes all files beginning with "._".

From feature request #82.