Removing ressource fork AppleDouble fails, bug #542
When the AFP client uses FPSetForkParms to set the ressource fork
length to 0, Netatalk will call unlink() in ad_rtruncate() in
libatalk/adouble/ad_write.c in order to to unlink the files
AppleDouble ressource fork.

The function ad_rtruncate() gets just the filename as path from the
caller, so obviously at this point the caller should have ensured that
it had called chdir() to the enclosing directory of the file.

Fixes bug #542.