libatalk/conf: re-generation of afp_voluuid.conf
In order to get rid of stale entries in the UUID db file, rewrite it
every time we start or the netatalk process gets a SIGHUP.

Otherwise we may leave orphaned entries for deleted volumes. If later
someone adds a share with the same name but a different path, we trip
over it when using MySQL CNID backend:

the problem happens because to look up the appropriate MySQL table, we
use the VolUUID found in afp_voluuid.conf, and then use that to look up
the information in the MySQL CNID database. afp_voluuid.conf stores a
volume name and UUID, but when you look up the table in the MySQL CNID
backend, the full volume path is stored. So if you look up the UUID in
afp_voluuid.conf, you get the UUID for the old volume that no longer
exists, and then when you try looking up the CNID database in MySQL, the
path does not match and you get the above error.

This essentially ensures deleted volumes are removed from
afp_voluuid.conf.

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: HAT <hat@fa2.so-net.ne.jp>