afpd: reading from file may fail
When we get a request to list xattrs of a file that:

o is an open fork file handle in use by the client

o has no associated metadata

we call ad_close() on the adouble handle which does a refcount decrement
which closes the file handle.

The fix is remembering whether we the ad_open() actually succeeded and
only call ad_close() in that case.

Bug: https://sourceforge.net/p/netatalk/bugs/619/

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Andrew Stormont <andyjstormont@gmail.com>
(cherry picked from commit b39b9388f854d3966c004c9c630c023f09c157e6)