New Global option "zeroconf name", FR#99
<https://sourceforge.net/p/netatalk/feature-requests/99/>

Stanislav Kozyrev said:

I was playing with Zeroconf the other day and noticed that the
Netatalk doesn't support specifying a service name. According to
Apple[1], 'For most hardware devices, the default service name should
be the full make and model of the product. For example, something like
"Apple MacBook Pro"'. In my particular case, this feature became
important because Finder had a problem to distinguish what service
type a user would like to use when the same hostname was advertising
both AFP and SMB services. In the end, I wanted to have something like
below:

; Global server settings
[Global]
    service name = MicroServer Gen8
    mimic model = Macmini4,1

I've failed to find anything on the Internet, so I tried to add a
missing functionality myself. With an attached patch applied and a
service name specified, AFP appears as a Mac Mini with a human
readable name and SMB is shown as a regular Windows box (one with a
bluescreen). If the service name is missing the Netatalk will fallback
to a host name as usual.

Please review the patch and let me know what you think. Thank you in
advance.

[1] See " 9. What should I pass in for the "name" parameter when
registering a service?" at
https://developer.apple.com/library/content/documentation/Cocoa/Conceptual/NetServices/Articles/faq.html

-----------------------------

replace:
service name -> zeroconf name

[Global]
    zeroconf name = MicroServer Gen8

-----------------------------

Signed-off-by: Stanislav Kozyrev <stanislav.kozyrev@gmail.com>
Reviewed-by: Andrew Stormont <andyjstormont@gmail.com>
Reviewed-by: HAT <hat@fa2.so-net.ne.jp>
Reviewed-by: Ralph Boehme <slow@samba.org>