- do not disable quota support if some of rpc/rpc.h, rpc/pmap_prot.h,   rpcsvc/rquota.h are missing.
- fix libquota test to check for the netbsd-6 release quota API, not
  the prerelease proplib-based API in 5.99 that got removed before
  netbsd-6.

This logic should probably be simplified, and expanded some to
simplify the mess of quota-related ifdefs in the source; the rquotad
support connected to the rpc headers and -lrpcsvc should be
independent of the local quota support, and can be probed for
independently. However, this needs to be coordinated with upstream,
not done in pkgsrc.