docs: manual: Remove wrong TCP-over-TCP info; minor copy editing
The main reason for this commit is to remove this text:

  Tunneling TCP over TCP (as SSH does) is not the best idea.
  There exist better solutions like VPNs based on the IP
  layer.

It's true that TCP-over-TCP is a bad idea, because it can lead
to TCP "meltdown"; however, it is *not* true that tunneling via
SSH implies TCP-over-TCP. Indeed, because a VPN is a much more
general solution to tunneling, it is likely to perform worse
than SSH for the simple case of connecting to a single, specific
AFP Server.

Consider the following port forwarding (using OpenSSH):

  ssh -N -L localhost:1234:localhost:548 example.com

When an AFP Client communicates with an AFP Server through this
tunnel, the data they exchange is encapsulated in TCP and decap-
sulated from TCP such that TCP is never encapsulated by TCP;
that is, TCP-over-TCP *never* occurs (view the following diagram
with a monospace font and with at least 93 coloumns of text):

                     localhost:1234           example.com:22           localhost:548
                           |                        |                        |
                           |                        |                        |
                           V   encrypt(AFP)--+      V                        V
                                             |
  +--------------+         +--------------+  V      +--------------+         +--------------+
  |          AFP | AFP/TCP |     AFP      | ???/TCP |     AFP      | AFP/TCP | AFP          |
  |   data <=====o=========o==============o=========o==============o=========o=====> data   |
  |              |         |              |         |              |         |              |
  |  AFP Client  |         |  SSH Client  |         |  SSH Server  |         |  AFP Server  |
  +--------------+         +--------------+         +--------------+         +--------------+

Anyway, while I was in the source rummaging around, I took the
opportunity to perform some copy editing, including a few mild
additions and clarifications.