libatalk: fix parsing of macOS created AppleDouble files
header_len is just AD_DATASZ_OSX which is 82. Pass the size of the AppleDouble
file to parse_entries() so the bound checks correctly work with the file size,
not just the header size.

With an FinderInfo AppleDouble entry that contains embedded xattrs, the
FinderInfo entry will be much larger then 32 bytes, typically it looks like
this:

Entry ID   : 00000009 : Finder Info
Offset     : 00000032 : 50
Length     : 00000EB0 : 3760

As offset + length is bigger then FinderInfo, parse_entries() fails the
validation.

Signed-off-by:    Ralph Boehme <slow@samba.org>
Reviewed-by:      Daniel Markstedt <daniel@mindani.net>