[2.3] papd - More cleanup and moderization of print_cups.c (#672)
* [2.3] papd- Use current IPP status and command names.

Replace obsolete IPP command names and status names with current versions.

* [2.3] papd - Update to cupsSetPassword2()

The function cupsSetPassword() is deprecated. Update all calls to cupsSetPassword2() and update the callback function to support the extra parameters passed.

* [2.3] papd - Remove unneeded IPP commands to get language info.

The ippNewRequest() function now automatically adds "attributes-charset" and "attributes-natural-language" to a request.

* [2.3] papd - Remove unneeded "language" definitions

* [2.3] papd - Minor edit

Set this to 'true' instead of '1'.

* [2.3] papd - Remove last ippNew() call

Missed one.