[2.3] atalkd: Fix for RTMP broadcast bug. (#680)
* [2.3] atalkd: Provisional fix for RTMP broadcast bug.

If the (only) interface was configured with the `-router` switch, automatically enable quirks mode. Otherwise don't. This is a hack as the split horizon logic is clearly broken here. If a single interface is configured as a router, it should ALWAYS broadcast RTMP packets no matter what.

* [2.3] atalkd: Update atalkd man page, remove 'quirks mode'

* [2.3] atalkd: Final fix for RTMP Broadcast

Simplified the conditional. Also removed the "quirks mode" option.