meson: Allow choosing shared or static libraries to build, GitHub #1305
Use the standard library() target for libatalk and UAM libraries, which makes the build system sensitive to the -Ddefault_library option to control whether to build static, shared, or both libraries.

The UAM libraries need the `b_lundef=false' workaround since they're not true shared libraries, but rely on symbols in netatalk proper.

Signed-off-by: Daniel Markstedt <daniel@mindani.net>