Initial AppleTalk fixes
Many changes to integrate several of the old AppleTalk daemons and utilities into the 3.x codebase. Most of it was dealing with changes to libatalk's unicode conversion interface.

The following now compile and run:
-atalkd service (tested working)
-papd service (tested working)
-timelord service (runs, but not tested further)
-a2boot service (runs, but not tested further)
-nbplkup (tested working)
-nbprgstr (tested working)
-nbpunrgstr (tested working)
-getzones (tested working)

Note that as of now, atalkd, papd, and getzones are hardcoded to use the MAC_ROMAN character encoding. This was already the case on FreeBSD and platforms using systemd as they had no way to set and use the systemwide character sets anyway.