testsuite: Don't treat initial spectest.sh run as a failure
The initial run of the spectest.sh script generates a config file that needs to be edited before running the whole suite. Before we treated this as a failed test to call attention this this fact. However this blocked CI jobs when tests and testsuite were both enabled and tests were ran.