testsuite: Consolidate spectest into a single suite
The testsuite grouping have been removed, and all spectests
are in a single suite. The tier 2 tests are enabled with
the -c flag. The sleep and readonly tests are enabled with
the -f flag.

Several tests have gotten improved setup and teardown
for better consistency.

We now run a testsuite job for each AFP version in
the GitHub actions CI.