macipgw: Default port value for zip/ddp service
On implementations relying on musl, getservbyname() only supports IP
protocols so calls to retrieve details for a DDP service fail, no
matter if "ddp" services are defined in /etc/services.

This change follows the pattern on atalkd and nbplkup where a
default port number is used for a DDP service unless getservbyname()
provides a value for it.

Signed-off-by: Antonio Pastor <antonio.pastor@gmail.com>