Protect Netatalk metadata EA
Currently, a client can create new files on the server with the same name as the Netatalk metadata extended attribute (org.netatalk.Metadata). If this EA has valid AppleDouble data (ex: from another Netatalk host), the server will treat the EA as read-only and not be able to update the data.

Additionally clients will not be able to update metadata on the file such as comments or FinderInfo. The only way to fix the problem is deleting the extended attribute on the server side, or deleting the file.