Convert documentation from XML to Markdown, GitHub #1588
Leveraged pandoc's docbook filter to convert to CommonMark,
then cmark-gfm (or cmark for man pages only) to transcode
to roff or html.

This gives us a much, much leaner toolchain for generating documentation
through the build system. No need to provide the heavy XML and Docbook suites.

Localization is done with po4a, which uses gettext in the backend.

Introduces with-docs, with-docs-l10n, and with-website Meson options.