meson: Introduce with-unicode-data option to build case tables
The new with-unicode-data Meson build system option will control whether to attempt to rebuild the Unicode case table sources or not. This is disabled by default, which is a change from before.

This makes it so that there are no surprising changes to local source files unless you deliberately ask for it.