Link with shared iniparser library, GitHub #1870
Remove the forked iniparser library from the codebase, and link with the standard iniparser shared library instead.

Tested working with iniparser 3.1 shipped with NetBDS and OpenBSD, as well as 4.x shipped with all other distros. Only on Solaris 11 are we lacking a packaged version of the library, so in the CI build I pull down the latest tarball from GitLab - 4.3.5 at the time of writing - and build it on the fly.

This involves changing some of the function calls and introducing a macro to get the same behavior as before. Notably, the section + key parameter is combined in the upstream library, while separated in our fork. So I had to do some workarounds. Our fork is a bit more elegant in my opinion.

In order to get linking to work on Solaris 11, I had to enable RPATH for the libatalk library as well. RPATH is only on by default for Solaris and NetBSD.

Resolves a number of build errors and static analysis bugs that were encountered in the process.

Signed-off-by: Daniel Marksted <daniel@mindani.net>