uams: Support for OpenBSD flavor crypt_checkpass() validation
The crypt() function is flagged as deprecated for OpenBSD,
so using the slightly more elegant crypt_checkpass() when available.

Consistently null out the buffer holding the password
when we're done using it.

In Meson, only link with crypt when the library is found.

Also refactors the crypt/shadow checks in Meson to give a
more accurate summary output.

Signed-off-by: Daniel Markstedt <daniel@mindani.net>