Apply additional hardening to the Netatalk Metadata EA handling
Right now a client can still delete the `org.netatalk.Metadata` extended attribute using `xattr -d`. Clients should have no access to this data at all as it is used exclusively by the Netatalk server for storing its metadata.

In addition, adjust `sys_set_ea()` to silently fail if a client attempts to copy a file with a `org.netatalk.Metadata` EA attached to it from a local machine to the server. Right now it will abort file copies if this EA is present and macOS does not provide a descriptive error as to why. This should fix the problems discussed in #2000.