meson: Option for controlling CUPS libdir and pap backend
The new -Dwith-cups-pap-backend boolean option, default true,
controls whether to install the pap backend for CUPS or not.
This is handy when you can't or won't install configuration files
for other libraries on the system.

In addition, the new -Dwith-cups-libdir-path which takes a path
defines the root path where CUPS library files are kept,
used to install the pap backend into a cups/backend subdir.