meson: Generate Unicode lookup table sources before using them
The build system will attempt to build the unicode static library
before generating some of the source code files used for the target.
This is normally not a problem since we bundle pre-built sources.
But in some packaging environments (notably Debian) we strip out
the pre-built sources and rely on the source generation logic.

This fix reverses the sequence of those two actions.