Create bstring README with redistribution notes and full credits
When Frank Lahm added this library to revision control in early 2010,
the license.txt file from the bstring tarball was not included,
and the project has never included the full text of Paul Hsieh's
BSD 3-clause license with the Netatalk source code.

Since bstrlib.txt in the original tarball says that the user
can choose to make the library "available" under either BSD or GPL v2,
I hereby make it explicit that the Netatalk project chooses the GPL v2.

Signed-off-by: Daniel Markstedt <daniel@mindani.net>