Handle DSIWrite() bug in ASC 3.7.x
When sending a FPWrite() packet with data length/reqcount of zero (programs like Stuff-It Expander 5.1.4 do this for some reason), some revisions of AppleShare Client (confirmed 3.7.2 to 3.7.4) also set the write offset in the DSI header to zero. This is blatantly wrong and must always be set to 12 to accommodate the packet header. Fixes #2078