meson: Do not install the pap CUPS backend by default
The pap backend for CUPS argualy modifies a 3rd party
library which is not netatalk or the core system.
Therefore, many OSes have policies against one package
modifying the configuration of another package.
Notable NixOS.

This change makes it so that the pap CUPS backend
has to be enabled explicitly at compile time to be installed.