docs: Reduce overlap between install chapter and install readme
The INSTALL.md file has gotten much better at covering
the basics of using git, meson, and understand dependencies.
Therefore I now remove some redundant information on this topic
from the Installation chapter in the html manual.

Conceptually, the Installation chapter is now meant to
give the high level overview, plus an in-depth discussion
on the purpose of each dependent library.

Conversely, INSTALL.md can focus on the mechanics of working with
the source code.